feat: create diff directories for agorithm and graphs
