+++ index.html
@@ -4,6 +4,14 @@
     <meta charset="UTF-8">
     <title>Title</title>
     <link rel="stylesheet" href="style.css">
+
+
+    <script src="graph/vertex_info.js" charset="utf-8"></script>
+    <script src="graph/vertex.js" charset="utf-8"></script>
+    <script src="algorithms/a_star.js" charset="utf-8"></script>
+    <script src="algorithms/dijkstra.js" charset="utf-8"></script>
+    <script src="algorithms/bfs.js" charset="utf-8"></script>
+    <script src="algorithms/dfs.js" charset="utf-8"></script>
     <script src="script.js" charset="utf-8"></script>
 </head>
 <body>
@@ -35,25 +43,31 @@
                 Statistic information
             </h1>
         </div>
-        <div class="bfs-info" id="bfs-info">
+        <div class="bfs-block">
             <h2>
                 BFS algorithm
             </h2>
+            <div class="bfs-info" id="bfs-info"></div>
         </div>
-        <div class="dfs-info" id="dfs-info">
+        <div class="dfs-block">
             <h2>
                 DFS algorithm
             </h2>
+            <div class="dfs-info" id="dfs-info">
+
+            </div>
         </div>
-        <div class="a-info" id="a-info">
+        <div class="a_star-block">
             <h2>
                 A-algorithm
             </h2>
+            <div class="a-info" id="a-info"></div>
         </div>
-        <div class="greedy-info" id="greedy-info">
+        <div class="greedy-block">
             <h2>
                 Greedy algorithm
             </h2>
+            <div class="greedy-info" id="greedy-info"></div>
         </div>
         <!--        </div>-->
         <!--        <div class="dfs-stat"></div>-->
