+++ style.scss
@@ -161,19 +161,19 @@ body {
       justify-content: center;
     }
 
-    .bfs-info {
+    .bfs-block {
       background-color: #4CAF50;
     }
 
-    .dfs-info {
+    .dfs-block {
       background-color: #888888;
     }
 
-    .a-info {
+    .a_star-block {
       background-color: #f44336;
     }
 
-    .greedy-info {
+    .greedy-block {
       background-color: darkorange;
     }
